backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
weaviate-client==3.26.2
python-dotenv==1.0.0
